# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers: 
  - url: 'http://localhost:8089/api'
info:
  description: API for warehouses, markets and its users
  version: 1.0.0
  title: Market_API
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Get users
      description: Get all users
      operationId: getUsers
      parameters:
        - name: order
          in: query
          description: Order results by name and id
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - id
        - name: name
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description:  Maximum number of results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: Offset of the results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
    post:
      tags:
        - users
      summary: Create a user
      description: Create a user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User to create
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
    put:
      tags:
        - users
      summary: Update user
      description: Update user
      operationId: updateUser
      parameters:
        - name: token
          in: query
          description: Token of the user
          required: true
          schema:
            type: string
        - name: user
          in: query
          description: User
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User updated
        '400':
          description: Bad request
        '404':
          description: User not found
  '/users/{userId}':
    get:
      tags:
        - users
      summary: Get a user
      description: Get a user
      operationId: getUser
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      tags:
        - users
      summary: Delete users
      description: Delete user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: token
          in: query
          description: Token of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
  '/users/token/{userId}':
    get:
      tags:
        - users
      summary: Get a users token
      description: Get a token
      operationId: getToken
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: User's name
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: User's password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  '/users/password/{userId}':
    get:
      tags:
        - users
      summary: Get a users password
      description: Get a password
      operationId: getPassword
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: User's name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Password found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
components: 
  schemas: 
...
    User:
      type: object
      required:
        - name
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: number
        password:
          type: string
        address:
          type: string
        token:
          type: string
      example:
        id: 'u2'
        name: 'Irene'
        password: 'tdlp15'
        address: 'calle 35ª'
        token: '2345678909654t32'